backend/requirements.txt
flask>=2.0.1,<3.0.0
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.27.0
werkzeug>=2.0.0,<3.0.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.0.0
huggingface-hub==0.8.1
torch==1.13.1
transformers==4.21.0 